Move implementation to sipmsgops
